# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: CloudCarbon API
  description: |
    Open-source multi-cloud GreenOps platform for unified cost and carbon optimisation.

    **Authentication**: All endpoints (except `/health`) require either:
    - A `Bearer` JWT access token in the `Authorization` header, or
    - A `cc_`-prefixed API key in the `X-API-Key` header.

    **Session status**: Session 5 (final) — agent infrastructure, five autonomous agents,
    notification delivery, Prometheus metrics, structured logging, rate limiting, and
    production Helm chart are complete.
  version: "0.5.0"
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.cloudcarbon.io
    description: Production (future)

# =============================================================================
# Security
# =============================================================================
security:
  - BearerAuth: []
  - ApiKeyAuth: []

# =============================================================================
# Tags
# =============================================================================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from /auth/token

    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key prefixed with `cc_`, obtained from the dashboard

  # ---------------------------------------------------------------------------
  # Shared Parameters
  # ---------------------------------------------------------------------------
...
  schemas:
...
    OverviewReportResponse:
      type: object
      properties:
        period:
          type: object
          properties:
            start:
              type: string
              format: date
            end:
              type: string
              format: date
        total_cost_usd:
          type: number
          format: double
        total_co2e_kg:
          type: number
          format: double
        scope3_co2e_kg:
          type: number
          format: double
        scope3_pct:
          type: number
          format: double
        total_water_litres:
          type: number
          format: double
        water_stress_adjusted_litres:
          type: number
          format: double
        carbon_efficiency:
          type: number
          format: double
          description: kg CO2e per $1000 spend
        cost_mom_pct:
          type: number
          format: double
          description: Month-over-month cost change (%)
        co2e_mom_pct:
          type: number
          format: double
          description: Month-over-month carbon change (%)
        water_mom_pct:
          type: number
          format: double
          description: Month-over-month water change (%)
        by_provider:
          type: array
          items:
            type: object
        by_service_category:
          type: array
          items:
            type: object
        top_resources_by_co2e:
          type: array
          items:
            type: object
    CarbonReportResponse:
      type: object
      properties:
        period:
          type: object
          properties:
            start:
              type: string
              format: date
            end:
              type: string
              format: date
            granularity:
              type: string
              enum: [daily, weekly, monthly]
        time_series:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date
              scope1:
                type: number
              scope2_location:
                type: number
              scope2_market:
                type: number
              scope3_total:
                type: number
              total:
                type: number
        by_region:
          type: array
          items:
            type: object
        scope3_breakdown:
          type: object
          properties:
            cat1_embodied:
              type: number
            cat3_upstream:
              type: number
            cat12_eol:
              type: number
            total:
              type: number
            confidence_distribution:
              type: object
    WaterReportResponse:
      type: object
      properties:
        period:
          type: object
        total_water_litres:
          type: number
          format: double
        total_water_stress_adjusted_litres:
          type: number
          format: double
        by_region:
          type: array
          items:
            type: object
        by_service_category:
          type: array
          items:
            type: object
        high_stress_regions:
          type: array
          items:
            type: object
...
paths:
...
  /reports/overview:
    get:
      tags: [Reports]
      summary: Multi-dimensional overview report
      description: |
        Return cost, carbon, and water totals with month-over-month changes,
        provider breakdown, service category breakdown, and top 10 resources by CO2e.
      operationId: overviewReport
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: provider
          in: query
          schema:
            type: string
        - name: mode
          in: query
          schema:
            type: string
            enum: [cost, sustainability]
            default: cost
      responses:
        "200":
          description: Overview report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OverviewReportResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /reports/carbon:
    get:
      tags: [Reports]
      summary: GHG Protocol carbon report
      description: |
        Return a GHG Protocol-aligned carbon report with time series (Scope 2 + Scope 3),
        regional breakdown, and Scope 3 category detail with confidence distribution.
      operationId: carbonReport
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: provider
          in: query
          schema:
            type: string
        - name: granularity
          in: query
          schema:
            type: string
            enum: [daily, weekly, monthly]
            default: daily
      responses:
        "200":
          description: Carbon report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CarbonReportResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /reports/water:
    get:
      tags: [Reports]
      summary: Water consumption report
      description: |
        Return water consumption with regional breakdown (WRI Aqueduct stress scores),
        service category breakdown, and high-stress region alerts.
      operationId: waterReport
      security:
        - BearerAuth: []
        - ApiKeyAuth: []
      parameters:
        - name: start_date
          in: query
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          schema:
            type: string
            format: date
        - name: provider
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Water report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WaterReportResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
...